+++ CANtactCore/src/org/cantact/ui/TransmitTopComponent.java
@@ -167,15 +167,15 @@ public final class TransmitTopComponent extends TopComponent {
         CanFrame txFrame = new CanFrame();
         txFrame.setId(Integer.decode(this.idField.getText()));
         txFrame.setDlc(Integer.decode(this.dlcComboBox.getSelectedItem().toString()));
-        int[] data = {
-            Integer.decode(this.byte0Field.getText()),
-            Integer.decode(this.byte1Field.getText()),
-            Integer.decode(this.byte2Field.getText()),
-            Integer.decode(this.byte3Field.getText()),
-            Integer.decode(this.byte4Field.getText()),
-            Integer.decode(this.byte5Field.getText()),
-            Integer.decode(this.byte6Field.getText()),
-            Integer.decode(this.byte7Field.getText())
+        byte[] data = {
+            Integer.decode(this.byte0Field.getText()).byteValue(),
+            Integer.decode(this.byte1Field.getText()).byteValue(),
+            Integer.decode(this.byte2Field.getText()).byteValue(),
+            Integer.decode(this.byte3Field.getText()).byteValue(),
+            Integer.decode(this.byte4Field.getText()).byteValue(),
+            Integer.decode(this.byte5Field.getText()).byteValue(),
+            Integer.decode(this.byte6Field.getText()).byteValue(),
+            Integer.decode(this.byte7Field.getText()).byteValue()
         };
         txFrame.setData(data);
         DeviceManager.transmit(txFrame);
